refactor(allocator): add `Allocator::new_fixed_size` method (#22084)
Add a method `new_fixed_size` to `Arena` and `Allocator`. This moves logic for creating fixed-size allocators out of `FixedSizeAllocatorPool` and into core `Arena` implementation.

This is starting point for making fixed-size allocators safer, managing their own memory, and having platform-specific behavior for creation and drop (moving towards fixing the Windows OOM bug in Oxlint and `napi/parser` raw transfer).
